Move data functions from wind_turbine to data
