+++ windpowerlib/__init__.py
@@ -3,8 +3,8 @@ __license__ = "MIT"
 __version__ = "0.2.1dev"
 
 from .wind_turbine import WindTurbine  # noqa: F401
-from .wind_turbine import get_turbine_types  # noqa: F401
-from .wind_turbine import create_power_curve  # noqa: F401
+from .data import get_turbine_types  # noqa: F401
+from .power_curves import create_power_curve  # noqa: F401
 from .wind_farm import WindFarm  # noqa: F401
 from .wind_turbine_cluster import WindTurbineCluster  # noqa: F401
 from .modelchain import ModelChain  # noqa: F401
